==> src/3-aim2/3-figure-scripts/0-run-figure-scripts.R <==



#source(paste0(here::here(),"/src/3-aim2/3-figure-scripts/0-tables-aim2.R"))
source(paste0(here::here(),"/src/3-aim2/3-figure-scripts/1-aim-2-primary-figures.R"))
source(paste0(here::here(),"/src/3-aim2/3-figure-scripts/2-aim-2-figures-diff.R"))
source(paste0(here::here(),"/src/3-aim2/3-figure-scripts/3-aim-2-pathogen-infection-figures.R"))
source(paste0(here::here(),"/src/3-aim2/3-figure-scripts/4-aim-2-pathogen-specific-heatmaps.R"))
source(paste0(here::here(),"/src/3-aim2/3-figure-scripts/5-aim-2-abundance-heatmaps.R"))
source(paste0(here::here(),"/src/3-aim2/3-figure-scripts/6-aim-2-figures-subgroup.R"))
source(paste0(here::here(),"/src/3-aim2/3-figure-scripts/7-aim-2-figures-subgroup-age.R"))
source(paste0(here::here(),"/src/3-aim2/3-figure-scripts/8-covariate-exposure-associations.R"))
source(paste0(here::here(),"/src/3-aim2/3-figure-scripts/9-unadjusted-adjusted-comparison-figure.R"))
source(paste0(here::here(),"/src/3-aim2/3-figure-scripts/10-aim-2-tmle-figures.R"))
source(paste0(here::here(),"/src/3-aim2/3-figure-scripts/12-aim2-tr-figures.R"))
source(paste0(here::here(),"/src/3-aim2/3-figure-scripts/13-aim-2-all-diarrhea.R"))
